+++ src/f64x4_.rs
@@ -24,6 +24,9 @@ impl_simd! {
     T = f64,
     N = 4,
     Simd = f64x4,
+    optional_type_x86_inner { X86Inner = __m256d },
+    optional_type_arm_inner {},
+    optional_type_wasm_inner {},
   }
 
   #[inline]