+++ src/i16x16_.rs
@@ -17,6 +17,9 @@ impl_simd! {
     T = i16,
     N = 16,
     Simd = i16x16,
+    optional_type_x86_inner { X86Inner = __m256i },
+    optional_type_arm_inner {},
+    optional_type_wasm_inner {},
   }
 
   #[inline]