+++ src/i64x8_.rs
@@ -17,6 +17,9 @@ impl_simd! {
     T = i64,
     N = 8,
     Simd = i64x8,
+    optional_type_x86_inner { X86Inner = __m512i },
+    optional_type_arm_inner {},
+    optional_type_wasm_inner {},
   }
 
   #[inline]